basic rebuild test
Add basic ct test for rebuilding trees in parallel and native mode.

Need to move forward with rebuild test for rebuilding the store.  To help with this a NotifyFun is added to the store prompt, to allow for an async reply so that the controller can update the status following a prompt.