RepairFun to take clock comparison
Rather than give the RepairFun just a bunch of keys thta have differences, give th eRepairfun both the BlueClock and the PinkClock for each difference.

For  intra-riak-cluster repair, this may be irrelevant (as the repair should be by read_repair).  Howver, for inter-cluster repair this cna be used ot avoid needlessly sending out of date data between clusters only for it to be rejected on the other side.