Add rehahs support in native mode
If the previous clock is undefined in native mode - a null tree change should be generated rather than the old object being looked up to compare (as the lookup could only be confusing).

There is some potential here to think of doing a dirty segments thing.  When a fold over segments is requested, should the trees be soft-locked to check for queue any updates for those segments - and a new hahs calculated as part of the fold.  This would allow for a proper rehash.  TODO.