Fetch Clocks and Dirty Segments
When we call fetch_clocks a new version of the segment hash can be calculated as part fo the fold.

so now when fetch_clocks is called for segments, whose segments (for the related IndexNs) are marked as dirty.  If nothing happens between the query being requested and completing (like a load is started or a change is made for that segment), the segment hahs is updated at the end of the fold.

This should mean that if a tree has got out of sync with a store - leading to false negative reports in aAE exchanges, those AAE exchanges will repair the broken bits of the tree, stopping the same false faults from continually repeating.

Note if a parallel keystore is out of sync (and this is the cause of the discrepancy), then the riak_kv_vnode:rehash called as part of repairing a key will resolve this issue.