Consistent use of NULL SubKey
the SubKey gets bubble dup in folds by Leveled, so need to consistently use a null subkey in kv_index_tictactree.

The null_subkey of <<>> is used, as the sub_key needs to be a binary or list to fit in with the leveled definition of valid keys.  Previously it used the leveled internla definition of a null key ('null'), but this might lead to unexpected behaviour - kv_index_tictactree should follow the external API