Break up queue load
When rebuilding many stores on a node concurrently, where the cluster is subject to large PUT loads, the load from the queue after the load from the vnode store is complete can be expensive.

It may result in a massive work backlog on the penciller (over 100 items has been seen), and a very large bookie's memory.  It may also take sufficient time for sync calls after the load has complete to time out.

Previously the load queue did not pause when requested by the backend store, due to fear of timeouts.  Now as each item on the queue is its own message, the load is fragmented (first into batches then into messages) and the pauses can be respected without fear of timeouts.