Tune timeouts
There have been instances of SYNC_TIMEOUT being hit when calling rebuild_trees on an AAE controller during ownership handoff.

These leads to worker crashes.  The rebuild_trees can wait if in the loading state - so why not treat the timeout the same, and simply retry.

Also if this is don't rebuild - don't even try the currentstatus check.