Only check presence on scheduled builds
If broken - speed up by not checking presence.  Also don't check presence when fetching clocks.

Presence check is part of anti-entropy, but it is to tackle long-term entropy (i.e. that not handled by read repair).  So the use in scheduled rebuilds only is satisfactory - and there should therefore be less processing required in other queries now.