Still need to check_presence on fetch_clocks_nval
If there is a rebuild, and the rebuild discovers a key is not present, then there will be a cache delta.

However, that cache delta will need to run check presence, otherwise it will not see the delta (and further more will "correct" the delta in the store).

This is a shame, as it uses potentially quite a bit more CPU (and requires more journal snapshots to be taken).  But it is unavoidable if we wish to recoer from broken journal files via AAE