Mas updateto29 (#94)
* Update rebar.config

For merge to 3.0

* Use release tag

* Allow for merging of partial and full exchnage types

Now full exchange types can have modified ranges for clock compares, not just partial exchange types

* Randomise sub-selections

Avoid getting the same selections each time when two exchanges run in parallel

* Temp commit for debug

In cases where clock_compare has fairly large lists ... appears to be incredibly slow

* Correct log info

* Revert "Temp commit for debug"

This reverts commit 207ac19e0d061dd67e0930b74da4567107c8cb35.

* Allow for filtering by bucket on fetch_clocks_nval

Also improve logging on close

* Allow for logging of purpose

to allow in the logs full-sync aae and intra-cluster aae to be easily distinguished

* Add new options to type

* Update aae_exchange.erl

* Include purpose in EX001 correctly

* Only check presence on scheduled builds

If broken - speed up by not checking presence.  Also don't check presence when fetching clocks.

Presence check is part of anti-entropy, but it is to tackle long-term entropy (i.e. that not handled by read repair).  So the use in scheduled rebuilds only is satisfactory - and there should therefore be less processing required in other queries now.

* Still need to check_presence on fetch_clocks_nval

If there is a rebuild, and the rebuild discovers a key is not present, then there will be a cache delta.

However, that cache delta will need to run check presence, otherwise it will not see the delta (and further more will "correct" the delta in the store).

This is a shame, as it uses potentially quite a bit more CPU (and requires more journal snapshots to be taken).  But it is unavoidable if we wish to recoer from broken journal files via AAE

* Prevent crash/start/stop/start issue

If the tree cache starts without a saved cache, then it shouldn't save its cache on close unless a rebuild has been completed

* Bring eqc into main build

* travis friendly?

* If tree is cached - everything is a worthwhile reduction

* Apply suggestions from code review

Co-authored-by: Thomas Arts <thomas.arts@quviq.com>

* Correct mismatch segments when < MaxResults

* Export info about responses

* Repoint leveled to develop-2.9

Co-authored-by: Thomas Arts <thomas.arts@quviq.com>