Mas i1824 corruptedobj (#106)
When rebuilding the store in parallel mode, if an object is corrupted so it cannot be loaded, the corruption can be caught and the object skipped. It is also possible to pass in a function to perform a specific action when the exception is detected (e.g. prompt read repair).

They got de-dup'd as part of aae_keystore:do_load/3 - but clearly the duplication here was a mistake.